Deploying to a function not set as "deploy to container image" will not work.
At the time of writing, the Consumption plan doesn't support the "deploy as a container image" option. The instructions at https://learn.microsoft.com/en-us/azure/azure-functions/functions-how-to-custom-container carefully call out to use a "Premium plan", but don't explain that this enables the "Deploy as a container image" option. If one chooses the Consumption plan then uses this GitHub Action, it will appear to deploy correctly, but the container will not start, and the user will be left with a zombie function.